requirements.txt
# Dependencies for Python controlled MediaWiki Docker image Installation
# Since 2021-06-08
# Docker SDK for Python
#https://pypi.org/project/docker/
# docker
# since 2016-06-19 
# https://github.com/gabrieldemarmiesse/python-on-whales
python-on-whales
# https://pypi.org/project/Jinja2/
Jinja2
#
# avoid ModuleNotFoundError: No module named 'six'
six
# Access to SQL database
# https://pypi.org/project/mysql-connector-python/
mysql-connector-python
# https://pypi.org/project/py-3rdparty-mediawiki/
py-3rdparty-mediawiki>=0.4.10
# https://pypi.org/project/pylodstorage/
pylodstorage>=0.0.29
# Beautiful Soup HTML parser
# https://pypi.org/project/beautifulsoup4/
beautifulsoup4